requirements.txt
fastapi
uvicorn
python-multipart
Pillow
ffmpeg-python